[Update] comment migration

+++ db/migrate/20221221122005_create_comments.rb
@@ -2,8 +2,8 @@ class CreateComments < ActiveRecord::Migration[6.1]
   def change
     create_table :comments do |t|
 
-      t.references :customer, type: :integer, null: false, foreign_key: true
-      t.references :post, type: :integer, null: false, foreign_key: true
+      t.references :customer, type: :bigint, null: false, foreign_key: true
+      t.references :post, type: :bigint, null: false, foreign_key: true
 
       t.timestamps
